Added application route to menu page
Added an application route to the menu page, now just need to figure
out what it should look like. I’m thinking we keep the photo up at the
top with the word Menu overlaid. The menu is going to be a list of
food. An affix will be in the right rail that allows a searcher to
toggle between the different menus.